test: tpm2_getpubek: fix rm error message
test_all.sh outputs the following:

EvictControl succ.
test_tpm2_evictcontrol.sh pass
rm: cannot remove 'ek_pub.out': No such file or directory
getpubek successfully!
test_tpm2_getpubek.sh pass

This is a superfolous failure that doesn't harm anything, but
check for the files existence before deleting it to remove
this misleading output.

Fixes #225

+++ test/system/test_tpm2_getpubek.sh
@@ -34,12 +34,14 @@ handle_ek=0x81010005
 ek_alg=0x001
 output_ek_pub=ek_pub.out
 
-rm $output_ek_pub
+if [ -f $output_ek_pub ]; then
+    rm $output_ek_pub
+fi
 
- tpm2_getpubek  -H $handle_ek -g $ek_alg -f $output_ek_pub 
+tpm2_getpubek  -H $handle_ek -g $ek_alg -f $output_ek_pub 
 if [ $? != 0 ] || [ ! -e $output_ek_pub ];then
-echo "getpubek fail, please check the environment or parameters!"
-exit 1
+    echo "getpubek fail, please check the environment or parameters!"
+    exit 1
 fi
 
 echo "getpubek successfully!"
